Evaluating a distribution

The function returns its arguments as a named list. The function is used inside the utility functions. It is transformed to an expression using parse and evaluated using eval. This ensures that in the case of an RPL with Bayesian priors, recursion is handled automatically. This significantly simplifies translating the utility function to lists of parameters to use when optimizing the designs. It is also less error prone.

N(mu, sigma)

Np(mu, sigma)

LN(mu, sigma)

LNp(mu, sigma)

TR(mu, sigma)

TRp(mu, sigma)

U(mu, sigma)

Up(mu, sigma)

Arguments

mu
A parameter indicating the mean or location of the distribution depending on whether it is a normal, log-normal, triangular or uniform, or it can be another call to N, LN, U or TR if the model is an RPL with a Bayesian prior.

sigma
A parameter indicating the SD or spread of the distribution or it can be another call to N, LN, U or TR.

Value

A list of parameters

Functions

N: The normal distribution

Np: The normal distribution when applied to a prior

LN: The log normal distribution

LNp: The log-normal distribution when applied to a prior

TR: The triangular distribution

TRp: The triangular distribution when applied to a prior

U: The uniform distribution

Up: The uniform distribution when applied to a prior

Examples

if (FALSE) {
eval(parse(text = "N(0, 1)"))
eval(parse(text = "LN(0, 1)"))
eval(parse(text = "TR(0, 1)"))
eval(parse(text = "U(0, 1)"))
}