S-error

Calculates a "lower bound" sample size to obtain theoretically significant parameter estimates under the assumption that the priors are correct.

calculate_s_error(design_vcov, p, all, significance)

Arguments

design_vcov

A variance-covariance matrix returned by derive_vcov or returned by an estimation routine. The matrix should be symmetrical and K-by-K

p

A vector of parameters, e.g. the named vector of priors. This is used for c- and s-errors. Default value is NULL

all

If `TRUE` return a K or K-1 vector with parameter specific error measures. Default is `FALSE`.

significance

A t-value corresponding to the desired level of significance. The default is significance at the 5 t-value of 1.96.

Value

A vector giving the "minimum" sample size for each parameter or a single number with the smallest sample size needed for all parameters to be theoretically significant.