Generate the full factorial

make_full_factorial is a wrapper around expand.grid and generates the full factorial given the supplied attributes. The attributes can either be specified directly by the user or extracted from the list of utility functions using parse_utility.

generate_full_factorial(attrs)

Arguments

attrs
A named list of attributes and their levels

Value

A matrix containing the full factorial

Details

The full factorial is often used as the starting point to generate a candidate set. Note that the full factorial will include unrealistic and completely dominated alternatives. It is therefore advised to use a subset of the full factorial as a candidate set. The user can call generate_full_factorial and create a subset that is passed to generate_design using the `candidate_set` parameter, or supply a set of restrictions through the list of design options `design_opt`.

The function is mainly used inside generate_design and appropriate restrictions are then put in place.

Examples

opts <- list(
  level_balance = FALSE
)
attrs <- list(
  a1 = 1:5,
  a2 = c(0, 1)
)

generate_full_factorial(attrs)

#>    a1 a2
#> 1   1  0
#> 2   2  0
#> 3   3  0
#> 4   4  0
#> 5   5  0
#> 6   1  1
#> 7   2  1
#> 8   3  1
#> 9   4  1
#> 10  5  1

V <- list(
  alt1 = "b_a1[0.1] * a1[1:5] + b_a2[-2] * a2[c(0, 1)]",
  alt2 = "b_a1      * a1      + b_a2     * a2"
)

attrs <- parse_utility(V, opts)$attrs
generate_full_factorial(attrs)

#>    a1 a2
#> 1   1  0
#> 2   2  0
#> 3   3  0
#> 4   4  0
#> 5   5  0
#> 6   1  1
#> 7   2  1
#> 8   3  1
#> 9   4  1
#> 10  5  1