Make a design candidate

The function is a call to switch and a wrapper around the design algorithms random, federov and rsc.

make_design_candidate(
  parsed_utility,
  candidate_set,
  current_design_candidate,
  opts,
  iter_with_no_imp,
  type
)

Arguments

parsed_utility

A list containing the elements of the parsed utility expression obtained from parse_utility.

candidate_set

A valid candidate set. This is only used for the 'random' and 'federov' algorithms.

current_design_candidate

A candidate for the RSC algorithm. To allow for marginal changes the candidate is passed back in for each iteration.

opts

A list of design options

iter_with_no_imp

The iteration number

type

One of the implemented types of optimization algorithms

Value

A data.frame with the design candidate in the 'wide' format. It is explicitly returned as a dataframe to use with model.matrix and to set up the environment. Reduces the number of calls to as.data.frame()

Details

The purpose of passing in utility is two-fold: 1) We can determine n_alts 2) We can extract the attribute names for each utility function to allow us to place the correct restrictions on the design candidate. Specifically, we restrict all levels of unavailable attributes to zero for alternatives where they do not feature. This is to ensure that we do not give weight when deriving the variance-covariance matrix using derive_vcov. Furthermore, the Xs are "sorted" using the order of the candidate set, which ensures that when we calculate the sum of the probabilities times X, the correct columns are added together. See derive_vcov.