Make random draws

A common interface to creating a variety of random draws used to simulate the log likelihood function

make_draws(n_ind, n_draws, n_dim, seed, type)

Arguments

n_ind

Number of individuals in your sample

n_draws

Number of draws per respondent

n_dim

Number of dimensions

seed

A seed to change the scrambling of the sobol sequence.

type

A character string

Value

A matrix of dimensions n_ind*n_draws x n_dim of standard uniform draws

Examples

n_ind <- 10
n_draws <- 5
n_dim <- 3

draws <- make_draws(n_ind, n_draws, n_dim, seed = 10, "scrambled-sobol")
head(draws)
#>           [,1]     [,2]     [,3]
#> [1,] 0.4163279 0.843947 0.599893
#> [2,] 0.5693955 0.708931 0.159586
#> [3,] 0.1755768 0.338082 0.778670
#> [4,] 0.6969957 0.439546 0.694511
#> [5,] 0.0600462 0.576020 0.368274
#> [6,] 0.9532566 0.992286 0.933824

draws <- make_draws(n_ind, n_draws, n_dim, seed = 10, "scrambled-halton")
head(draws)
#>       [,1]     [,2] [,3]
#> [1,] 0.500 0.666667 0.60
#> [2,] 0.250 0.333333 0.20
#> [3,] 0.750 0.222222 0.80
#> [4,] 0.125 0.888889 0.40
#> [5,] 0.625 0.555556 0.12
#> [6,] 0.375 0.111111 0.72