Evaluating a distribution

The function returns its arguments as a named list. The function is used inside the utility functions. It is transformed to an expression using parse and evaluated using eval. This ensures that in the case of an RPL with Bayesian priors, recursion is handled automatically. This significantly simplifies translating the utility function to lists of parameters to use when optimizing the designs. It is also less error prone.

normal(mu, sigma)

normal_p(mu, sigma)

lognormal(mu, sigma)

lognormal_p(mu, sigma)

triangular(mu, sigma)

triangular_p(mu, sigma)

uniform(mu, sigma)

uniform_p(mu, sigma)

Arguments

mu

A parameter indicating the mean or location of the distribution depending on whether it is a normal, log-normal, triangular or uniform, or it can be another call to normal, lognormal, uniform or triangular if the model is an RPL with a Bayesian prior.

sigma

A parameter indicating the SD or spread of the distribution or it can be another call to normal, lognormal, uniform or triangular.

Value

A list of parameters

Functions

normal(): The normal distribution

normal_p(): The normal distribution when applied to a prior

lognormal(): The log normal distribution

lognormal_p(): The log-normal distribution when applied to a prior

triangular(): The triangular distribution

triangular_p(): The triangular distribution when applied to a prior

uniform(): The uniform distribution

uniform_p(): The uniform distribution when applied to a prior