Compute the radical inverse

Equation 2 in Bhat (2003)

radical_inverse(n_dim, primes, count, digit, perms)

Arguments

n_dim

Number of dimensions

primes

A vector of prime numbers

count

A matrix

digit

A vector

perms

A matrix of the permutations. Defaults to a set of Braaten-Weller permutations.

References

Bhat, C. n_draws., 2003, Simulation Estimation of Mixed Descrete Choice Models Using Randomized and Scrambled Halton Sequences, Transportation Research Part B, 9, pp. 837-855