Make a design candidate

The function is a call to switch and a wrapper around the design algorithms random, federov and rsc.

make_design_candidate(candidate_set, opts, utility, type)

Arguments

candidate_set
A valid candidate set

opts
A list of design options

utility
The list of utility expressions

type
One of the implemented types of optimization algorithms

Value

A list the length of nalts with the design matrix for each alternative

Details

The purpose of passing in utility is two-fold: 1) We can determine nalts 2) We can extract the attribute names for each utility function to allow us to place the correct restrictions on the design candidate. Specifically, we restrict all levels of unavailable attributes to zero for alternatives where they do not feature. This is to ensure that we do not give weight when deriving the variance-covariance matrix using derive_vcov. Furthermore, the Xs are "sorted" using the order of the candidate set, which ensures that when we calculate the sum of the probabilities times X, the correct columns are added together. See derive_vcov.